Removed project setup errors

+++ passetup.sh
@@ -14,7 +14,14 @@ then
   exit 1
 fi
 
+if [ ! -s ./frontend/.env.example ]
+then
+  echo "ERROR: File ./frontend/.env.example does not exist or empty."
+  exit 1
+fi
+
 sed "s/MYSQL_ROOT_PASSWORD:.*[[:space:]]*/MYSQL_ROOT_PASSWORD: $passwrd/" ./docker-compose.yml.example > ./docker-compose.yml
 sed "s/DB_PASSWORD=.*[[:space:]]*/DB_PASSWORD=$passwrd/" ./backend/.env.example > ./backend/.env
+cp frontend/.env.example frontend/.env
 
 exit 0
